Move uploadFileOnReference to reference module.
