+++ src/domain/storage/storage-bucket/storage.bucket.module.ts
@@ -10,7 +10,6 @@ import { StorageBucketService } from './storage.bucket.service';
 import { StorageBucketAuthorizationService } from './storage.bucket.service.authorization';
 import { VisualModule } from '@domain/common/visual/visual.module';
 import { Document } from '../document/document.entity';
-import { ReferenceModule } from '@domain/common/reference/reference.module';
 import { Profile } from '@domain/common/profile/profile.entity';
 import { UrlGeneratorModule } from '@services/infrastructure/url-generator';
 
@@ -20,7 +19,6 @@ import { UrlGeneratorModule } from '@services/infrastructure/url-generator';
     AuthorizationModule,
     AuthorizationPolicyModule,
     VisualModule,
-    ReferenceModule,
     UrlGeneratorModule,
     TypeOrmModule.forFeature([StorageBucket]),
     TypeOrmModule.forFeature([Document]),
